requirements.txt
alpaca-trade-api==0.49.1
appnope==0.1.0
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
idna==2.10
ipdb==0.13.3
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
numpy==1.19.1
pandas==1.1.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
websocket-client==0.57.0
websockets==8.1
